imported font, fixed navbar links
